+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.5.2] - 06-01-21
+## [0.6.0] - 06-01-21
 ### Added
 - API to decrypt value and blinding factor from a crossover
 - Note::new is not part of the public API